Refactor domain and repository files into subpackages
