+++ demo/src/main/java/com/example/demo/UserController.java
@@ -1,5 +1,7 @@
 package com.example.demo;
 
+import com.example.demo.domain.ZetchUser;
+import com.example.demo.repo.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
