--- demo/src/main/java/com/example/demo/UserRepository.java
@@ -1,8 +1,0 @@
-package com.example.demo;
-
-import org.springframework.data.repository.CrudRepository;
-
-// This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
-// CRUD refers Create, Read, Update, Delete
-public interface UserRepository extends CrudRepository<ZetchUser, Integer> {
-}
