+++ demo/src/main/java/com/example/demo/repo/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.demo.repo;
+
+import com.example.demo.domain.ZetchUser;
+import org.springframework.data.repository.CrudRepository;
+
+/**
+ * User Repository Interface
+ * Will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
+ */
+public interface UserRepository extends CrudRepository<ZetchUser, Integer> {
+}
